Userlist and Whisper
Now there's a list of other users in chat, and doubleclicking on one trigger the whisper function, which will be sent only to that user and appear in italcs. Also, users that haven't pinged in over a set timeout are marked offline, and if longer, removed entirely. This might not play nice because of the ping behavior, may need to keep temp trashcan